requirements.txt
streamlit==1.55.0
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.23
langchain-google-genai==4.2.1
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
chromadb==1.5.5
sentence-transformers==5.3.0
google-genai==1.69.0
google-generativeai==0.8.6
python-dotenv==1.2.2